ext.img_display: add sanity checks to quit(), fixes #347

+++ ranger/ext/img_display.py
@@ -131,7 +131,8 @@ class W3MImageDisplayer(ImageDisplayer):
                 filename = path)
 
     def quit(self):
-        if self.is_initialized:
+        if self.is_initialized and self.process \
+                and self.process.poll() is not None:
             self.process.kill()
 
 class ITerm2ImageDisplayer(ImageDisplayer, FileManagerAware):
